+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Migrated core, temp, ui, world and ship systems to root systems.rs file for better overview
 - Move set_window_icon temporary system to PreStartup schedule
+- Move configure_physics_engine to GameCreate state
 
 ## [0.0.19] - 2023-10-31